# public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Sabek APis
  version: 1.0.0
  title: Sabek Frontend API's
basePath: /api/
...
paths:
...
  /carts/getDeletedCartItems:            
    post:
      tags:
        - Delete cart item
      summary: Get deleted items
      description: 'Get deleted cart items'
      operationId: getDeletedCartItems
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request  
...
securityDefinitions:
  api_token:
    type: apiKey
    name: apiToken
    in: header